v1.01
what's new:
--added features for the case when acknowledge CLS is not receiving
correctly by the client (when dmg and lost bit of server are set to 1,
this could happen). It detects if the base reaches the last packet needs
to send, and breaks the loop once time is out, disregarding on if the
CLS is received. This mechanism works, but not theoretically right, coz
it has small probability that last packet still not successfully
received by the server  till timeout.